requirements.txt
flask
flask-restx
requests
werkzeug
gunicorn
pytest
Flask-SQLAlchemy
Pillow